Extract the shared carousel core from the two tools-page carousels
