+++ src/components/CategoryPanelsCarousel/styles.module.css
@@ -1,61 +1,7 @@
-/* Template layout: the section header and the outlined circular arrows
-   share the top row, the scroller runs full width beneath them. */
-.carousel {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.headerRow {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 1.5rem;
-}
-
-.headerContent {
-  display: flex;
-  align-items: baseline;
-  gap: 0.75rem;
-  flex-wrap: wrap;
-  min-width: 0;
-}
-
-.arrows {
-  display: flex;
-  align-items: center;
-  gap: 0.6rem;
-  flex-shrink: 0;
-}
-
-.scroller {
-  list-style: none;
-  margin: 0;
-  padding: 0.25rem;
-  display: flex;
-  gap: 1.25rem;
-  overflow-x: auto;
-  overscroll-behavior-x: contain;
-  scroll-snap-type: x proximity;
-  scroll-behavior: smooth;
-  scrollbar-width: none;
-  -ms-overflow-style: none;
-  min-width: 0;
-}
-
-.scroller::-webkit-scrollbar {
-  display: none;
-}
-
 /* The template's 393px category card. */
 .item {
   flex: 0 0 min(393px, 85vw);
   max-width: 393px;
-  scroll-snap-align: start;
-}
-
-.item > * {
-  height: 100%;
 }
 
 /* The template's category card: a Clarity panel on the Neutral Grey
@@ -146,122 +92,7 @@ html[data-theme="dark"] .seeAll:hover {
 /* Rows sit bare on the card; their compact form in AppRow carries the
    negative-margin hover wash. */
 
-/* The template's outlined circular arrows */
-.arrow {
-  width: 48px;
-  height: 48px;
-  border-radius: 50%;
-  border: 1px solid rgb(var(--site-color-navy-rgb) / 0.5);
-  background: transparent;
-  color: var(--site-color-navy);
-  cursor: pointer;
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  transition: border-color 150ms ease, color 150ms ease, opacity 150ms ease,
-    box-shadow 150ms ease;
-  flex-shrink: 0;
-}
-
-html[data-theme="dark"] .arrow {
-  border-color: rgb(var(--site-color-card-text-rgb) / 0.4);
-  color: var(--site-color-cream);
-}
-
-.arrow:hover:not(:disabled) {
-  border-color: var(--site-color-brand-blue);
-  color: var(--ifm-font-color-base);
-  box-shadow: 0 1px 6px rgb(var(--site-color-navy-rgb) / 0.06);
-}
-
-html[data-theme="dark"] .arrow:hover:not(:disabled) {
-  border-color: var(--ifm-color-primary);
-}
-
-.arrow:focus-visible {
-  outline: 2px solid var(--site-color-navy);
-  outline-offset: 2px;
-}
-
-html[data-theme="dark"] .arrow:focus-visible {
-  outline-color: var(--site-color-amber);
-}
-
-.arrow:disabled {
-  opacity: 0.35;
-  cursor: not-allowed;
-}
-
-.arrow:disabled:hover {
-  border-color: rgb(var(--site-color-navy-rgb) / 0.2);
-  color: var(--site-color-text-muted);
-  box-shadow: none;
-}
-
-html[data-theme="dark"] .arrow:disabled:hover {
-  border-color: rgb(var(--site-color-card-text-rgb) / 0.25);
-}
-
-.dots {
-  display: none;
-}
-
 @media (max-width: 600px) {
-  .arrows {
-    display: none;
-  }
-
-  .scroller {
-    margin-left: -1rem;
-    margin-right: -1rem;
-    padding-left: 1rem;
-    padding-right: 1rem;
-    -webkit-mask-image: linear-gradient(
-      to right,
-      black 0,
-      black calc(100% - 28px),
-      transparent
-    );
-    mask-image: linear-gradient(
-      to right,
-      black 0,
-      black calc(100% - 28px),
-      transparent
-    );
-  }
-
-  .scroller.canPrev:not(.canNext) {
-    -webkit-mask-image: linear-gradient(
-      to right,
-      transparent,
-      black 28px,
-      black 100%
-    );
-    mask-image: linear-gradient(to right, transparent, black 28px, black 100%);
-  }
-
-  .scroller.canPrev.canNext {
-    -webkit-mask-image: linear-gradient(
-      to right,
-      transparent,
-      black 28px,
-      black calc(100% - 28px),
-      transparent
-    );
-    mask-image: linear-gradient(
-      to right,
-      transparent,
-      black 28px,
-      black calc(100% - 28px),
-      transparent
-    );
-  }
-
-  .scroller:not(.canPrev):not(.canNext) {
-    -webkit-mask-image: none;
-    mask-image: none;
-  }
-
   .item {
     flex: 0 0 300px;
     max-width: 300px;
@@ -269,33 +100,5 @@ html[data-theme="dark"] .arrow:disabled:hover {
 
   .panel {
     padding: 1.5rem 1.5rem;
-  }
-
-  .dots {
-    display: flex;
-    justify-content: center;
-    gap: 6px;
-    margin-top: 0.75rem;
-  }
-
-  .dot {
-    width: 6px;
-    height: 6px;
-    border-radius: 50%;
-    background: rgb(var(--site-color-navy-rgb) / 0.2);
-    transition: background-color 150ms ease, transform 150ms ease;
-  }
-
-  html[data-theme="dark"] .dot {
-    background: rgb(var(--site-color-card-text-rgb) / 0.25);
-  }
-
-  .dot.dotActive {
-    background: var(--site-color-brand-blue);
-    transform: scale(1.3);
-  }
-
-  html[data-theme="dark"] .dot.dotActive {
-    background: var(--ifm-color-primary);
   }
 }
